feat: add geode domain and application_name partitiion
