+++ deploy.bash
@@ -1,4 +1,5 @@
-VERSION="v4"
+PROJECT_NAME="geode-analytics"
+VERSION="v1"
 
 if [ -z $BRANCH_NAME ]; then
     # Jenkins runs script on git branch in a detached HEAD state.
@@ -40,31 +41,32 @@ if $IS_CHINA; then
     AWS_REGION="cn-north-1" # The region is the same in dev and prod environment
 fi
 
-DIST_OUTPUT_BUCKET="analytics-output-bucket"
-STACK_NAME="analytics-$ENVIRONMENT"
+DIST_OUTPUT_BUCKET="$PROJECT_NAME-output-bucket"
+STACK_NAME="$PROJECT_NAME-$ENVIRONMENT"
 
-# Run following commands only the first time to create bucket.
-# aws s3 mb s3://$DIST_OUTPUT_BUCKET --region $AWS_REGION
+# Run following command only the first time to create output bucket.
+aws s3 mb s3://$DIST_OUTPUT_BUCKET-$AWS_REGION --region $AWS_REGION
 
 cd ./deployment
 
 # Build project (Templates + Lambdas)
-# ./build-s3-dist.sh $DIST_OUTPUT_BUCKET analytics/$ENVIRONMENT $VERSION
+./build-s3-dist.sh $DIST_OUTPUT_BUCKET analytics/$ENVIRONMENT $VERSION
 
-# # Store Global Assets to S3 (Templates)
-# aws s3 cp ./global-s3-assets s3://$DIST_OUTPUT_BUCKET-$AWS_REGION/analytics/$ENVIRONMENT/$VERSION --recursive --acl bucket-owner-full-control
+# Store Global Assets to S3 (Templates)
+aws s3 cp ./global-s3-assets s3://$DIST_OUTPUT_BUCKET-$AWS_REGION/analytics/$ENVIRONMENT/$VERSION --recursive --acl bucket-owner-full-control
 
-# # Store Regional Assets to S3 (Lambdas)
-# aws s3 cp ./regional-s3-assets s3://$DIST_OUTPUT_BUCKET-$AWS_REGION/analytics/$ENVIRONMENT/$VERSION --recursive --acl bucket-owner-full-control
+# Store Regional Assets to S3 (Lambdas)
+aws s3 cp ./regional-s3-assets s3://$DIST_OUTPUT_BUCKET-$AWS_REGION/analytics/$ENVIRONMENT/$VERSION --recursive --acl bucket-owner-full-control
 
-# Deploy Backoffce Remote Config API Gateway (Zappa)
-./deploy-analytics-backoffice.sh $ENVIRONMENT $AWS_REGION
+# # Deploy Backoffce Remote Config API Gateway (Zappa)
+# ./deploy-analytics-backoffice.sh $ENVIRONMENT $AWS_REGION $PROJECT_NAME
 
-# # Deploy CloudFormation by creating/updating Stack
-# aws cloudformation deploy \
-#     --template-file ./global-s3-assets/game-analytics-pipeline.template \
-#     --stack-name $STACK_NAME \
-#     --capabilities CAPABILITY_IAM \
-#     --s3-bucket $DIST_OUTPUT_BUCKET-$AWS_REGION \
-#     --s3-prefix templates \
-#     $PARAMETER_OVERRIDES
+# Deploy CloudFormation by creating/updating Stack
+aws cloudformation deploy \
+    --template-file ./global-s3-assets/game-analytics-pipeline.template \
+    --stack-name $STACK_NAME \
+    --capabilities CAPABILITY_IAM \
+    --s3-bucket $DIST_OUTPUT_BUCKET-$AWS_REGION \
+    --s3-prefix templates \
+    --capabilities CAPABILITY_NAMED_IAM \
+    $PARAMETER_OVERRIDES
