+++ source/services/api/remote-configs/package.json
@@ -1,0 +1,23 @@
+{
+    "name": "remote-configs",
+    "description": "The remote-configs processor microservice for the game analytics pipeline solution",
+    "main": "index.js",
+    "version": "0.0.1",
+    "private": true,
+    "dependencies": {
+        "aws-sdk": "*",
+        "node-cache": "^5.1.2"
+    },
+    "devDependencies": {
+        "aws-sdk": "*"
+    },
+    "scripts": {
+        "build:init": "rm -rf package-lock.json && rm -rf dist && rm -rf node_modules",
+        "build:zip": "zip -rq remote-configs.zip .",
+        "build:dist": "mkdir dist && mv remote-configs.zip dist/",
+        "build": "npm run build:init && npm install --production && npm run build:zip && npm run build:dist"
+    },
+    "bundledDependencies": [
+        "aws-sdk"
+    ]
+}
